requirements.txt
asgiref==3.8.1
Django==5.2.3
paho-mqtt==2.1.0
psycopg2-binary==2.9.10
sqlparse==0.5.3